start of framework cleanup
